Refactored code and other duplication of original code with new ResolvedChannelResolver class
