+++ atlas-api/src/main/java/org/atlasapi/output/annotation/NextBroadcastAnnotation.java
@@ -1,64 +1,52 @@
 package org.atlasapi.output.annotation;
 
-import java.io.IOException;
-import java.util.List;
-import java.util.Set;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
-import java.util.stream.StreamSupport;
-
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.Iterables;
-import com.google.common.util.concurrent.Futures;
+import com.google.common.collect.ImmutableSet;
+import com.google.common.collect.ImmutableSet.Builder;
+import com.metabroadcast.common.ids.NumberToShortStringCodec;
 import com.metabroadcast.common.stream.MoreCollectors;
-import org.atlasapi.channel.Channel;
-import org.atlasapi.channel.ChannelEquivRef;
-import org.atlasapi.channel.ChannelResolver;
+import com.metabroadcast.common.time.Clock;
 import org.atlasapi.channel.ResolvedChannel;
 import org.atlasapi.content.Broadcast;
 import org.atlasapi.content.Content;
 import org.atlasapi.content.Item;
 import org.atlasapi.content.ResolvedBroadcast;
 import org.atlasapi.entity.Id;
-import org.atlasapi.entity.ResourceRef;
 import org.atlasapi.output.FieldWriter;
 import org.atlasapi.output.OutputContext;
+import org.atlasapi.output.ResolvedChannelResolver;
 import org.atlasapi.output.writers.BroadcastWriter;
-
-import com.metabroadcast.common.ids.NumberToShortStringCodec;
-import com.metabroadcast.common.time.Clock;
-
-import com.google.common.collect.ImmutableSet;
-import com.google.common.collect.ImmutableSet.Builder;
 import org.joda.time.DateTime;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.annotation.Nullable;
+import java.io.IOException;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
 public class NextBroadcastAnnotation extends OutputAnnotation<Content> {
 
     private static final Logger log = LoggerFactory.getLogger(NextBroadcastAnnotation.class);
     private final BroadcastWriter broadcastWriter;
     private final Clock clock;
-    private final ChannelResolver channelResolver;
+    private final ResolvedChannelResolver resolvedChannelResolver;
 
     private NextBroadcastAnnotation(
             Clock clock,
             BroadcastWriter broadcastWriter,
-            ChannelResolver channelResolver
+            ResolvedChannelResolver resolvedChannelResolver
     ) {
         super();
         this.clock = clock;
         this.broadcastWriter = broadcastWriter;
-        this.channelResolver = channelResolver;
+        this.resolvedChannelResolver = resolvedChannelResolver;
     }
 
     public static NextBroadcastAnnotation create(
             Clock clock,
             NumberToShortStringCodec codec,
-            ChannelResolver channelResolver
+            ResolvedChannelResolver resolvedChannelResolver
     ) {
         return new NextBroadcastAnnotation(
                 clock,
@@ -67,7 +55,7 @@ public class NextBroadcastAnnotation extends OutputAnnotation<Content> {
                         "broadcast",
                         codec
                 ),
-                channelResolver
+                resolvedChannelResolver
         );
     }
 
@@ -85,9 +73,15 @@ public class NextBroadcastAnnotation extends OutputAnnotation<Content> {
                 .filter(Broadcast::isActivelyPublished)
                 .collect(MoreCollectors.toImmutableList());
 
-        List<ResolvedBroadcast> resolvedBroadcasts = StreamSupport.stream(
-                nextBroadcast(broadcasts).spliterator(), false)
-                .map(broadcast -> ResolvedBroadcast.create(broadcast, resolveChannel(broadcast)))
+        List<Broadcast> nextBroadcasts = ImmutableList.copyOf(nextBroadcast(broadcasts));
+
+        Set<Id> channelIds = nextBroadcasts.stream()
+                .map(Broadcast::getChannelId)
+                .collect(MoreCollectors.toImmutableSet());
+        Map<Id, ResolvedChannel> channelMap = resolvedChannelResolver.resolveChannelMap(channelIds);
+
+        List<ResolvedBroadcast> resolvedBroadcasts = nextBroadcasts.stream()
+                .map(broadcast -> ResolvedBroadcast.create(broadcast, channelMap.get(broadcast.getChannelId())))
                 .collect(MoreCollectors.toImmutableList());
 
         writer.writeList(
@@ -112,46 +106,5 @@ public class NextBroadcastAnnotation extends OutputAnnotation<Content> {
             }
         }
         return filteredBroadcasts.build();
-    }
-
-    //TODO: Move resolution logic to query executor
-    private ResolvedChannel resolveChannel(Broadcast broadcast) {
-
-        try {
-            Channel channel = Futures.getChecked(
-                    channelResolver.resolveIds(
-                            ImmutableList.of(broadcast.getChannelId())
-                    ),
-                    IOException.class
-            )
-                    .getResources()
-                    .first()
-                    .orNull();
-
-            return ResolvedChannel.builder()
-                    .withChannel(channel)
-                    .withResolvedEquivalents(resolveEquivalents(channel.getSameAs()))
-                    .build();
-
-        } catch (IOException e) {
-            log.error("Failed to resolve channel: {}", broadcast.getChannelId(), e);
-            return null;
-        }
-
-    }
-
-    @Nullable
-    private Iterable<Channel> resolveEquivalents(Set<ChannelEquivRef> channelRefs) {
-        try {
-            if (channelRefs != null && !channelRefs.isEmpty()) {
-                Iterable<Id> ids = Iterables.transform(channelRefs, ResourceRef::getId);
-                return channelResolver.resolveIds(ids).get(1, TimeUnit.MINUTES).getResources();
-            }
-
-            return null;
-        } catch (InterruptedException | ExecutionException | TimeoutException e) {
-            log.error("Failed to resolve channel equivlents", e);
-            return null;
-        }
     }
 }
