+++ atlas-api/src/main/java/org/atlasapi/output/annotation/UpcomingBroadcastsAnnotation.java
@@ -1,21 +1,9 @@
 package org.atlasapi.output.annotation;
 
-import java.io.IOException;
-import java.util.List;
-import java.util.Set;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
-import java.util.stream.Stream;
-import java.util.stream.StreamSupport;
-
-import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Lists;
-import com.google.common.util.concurrent.Futures;
-import org.atlasapi.channel.Channel;
-import org.atlasapi.channel.ChannelEquivRef;
-import org.atlasapi.channel.ChannelResolver;
+import com.metabroadcast.common.ids.NumberToShortStringCodec;
+import com.metabroadcast.common.stream.MoreCollectors;
 import org.atlasapi.channel.Region;
 import org.atlasapi.channel.ResolvedChannel;
 import org.atlasapi.content.Broadcast;
@@ -24,44 +12,39 @@ import org.atlasapi.content.Content;
 import org.atlasapi.content.Item;
 import org.atlasapi.content.ResolvedBroadcast;
 import org.atlasapi.entity.Id;
-import org.atlasapi.entity.ResourceRef;
 import org.atlasapi.output.FieldWriter;
 import org.atlasapi.output.OutputContext;
+import org.atlasapi.output.ResolvedChannelResolver;
 import org.atlasapi.output.writers.BroadcastWriter;
-
-import com.metabroadcast.common.ids.NumberToShortStringCodec;
-import com.metabroadcast.common.stream.MoreCollectors;
-
 import org.joda.time.DateTime;
 import org.joda.time.DateTimeZone;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import javax.annotation.Nullable;
+import java.io.IOException;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
 /**
  * Marker for {@link BroadcastsAnnotation} to only write broadcasts that have not yet broadcast.
  */
 public class UpcomingBroadcastsAnnotation extends OutputAnnotation<Content> {
 
-    private static final Logger log = LoggerFactory.getLogger(UpcomingBroadcastsAnnotation.class);
-
     private final BroadcastWriter broadcastWriter;
     private final ChannelsBroadcastFilter channelsBroadcastFilter;
-    private final ChannelResolver channelResolver;
+    private final ResolvedChannelResolver resolvedChannelResolver;
 
     private UpcomingBroadcastsAnnotation(
             BroadcastWriter broadcastWriter,
-            ChannelResolver channelResolver
+            ResolvedChannelResolver resolvedChannelResolver
     ) {
         this.broadcastWriter = broadcastWriter;
         this.channelsBroadcastFilter = ChannelsBroadcastFilter.create();
-        this.channelResolver = channelResolver;
+        this.resolvedChannelResolver = resolvedChannelResolver;
     }
 
     public static UpcomingBroadcastsAnnotation create(
             NumberToShortStringCodec codec,
-            ChannelResolver channelResolver
+            ResolvedChannelResolver resolvedChannelResolver
     ) {
         return new UpcomingBroadcastsAnnotation(
                 BroadcastWriter.create(
@@ -69,7 +52,7 @@ public class UpcomingBroadcastsAnnotation extends OutputAnnotation<Content> {
                         "broadcast",
                         codec
                 ),
-                channelResolver
+                resolvedChannelResolver
         );
     }
 
@@ -77,9 +60,10 @@ public class UpcomingBroadcastsAnnotation extends OutputAnnotation<Content> {
     public void write(Content entity, FieldWriter writer, OutputContext ctxt) throws IOException {
         if (entity instanceof Item) {
             Item item = (Item) entity;
-            Stream<Broadcast> broadcastStream = item.getBroadcasts().stream()
+            List<Broadcast> filteredBroadcasts = item.getBroadcasts().stream()
                     .filter(Broadcast::isActivelyPublished)
-                    .filter(b -> b.getTransmissionTime().isAfter(DateTime.now(DateTimeZone.UTC)));
+                    .filter(b -> b.getTransmissionTime().isAfter(DateTime.now(DateTimeZone.UTC)))
+                    .collect(MoreCollectors.toImmutableList());
 
             if (ctxt.getRegions().isPresent()) {
                 List<Region> regions = ctxt.getRegions().get();
@@ -87,14 +71,19 @@ public class UpcomingBroadcastsAnnotation extends OutputAnnotation<Content> {
                 List<Broadcast> broadcasts = Lists.newArrayList();
                 regions.forEach(region -> {
                     Iterable<Broadcast> broadcastsToAdd = channelsBroadcastFilter.sortAndFilter(
-                            broadcastStream.collect(MoreCollectors.toImmutableList()),
+                            filteredBroadcasts,
                             region
                     );
                     Iterables.addAll(broadcasts, broadcastsToAdd);
                 });
 
+                Set<Id> channelIds = broadcasts.stream()
+                        .map(Broadcast::getChannelId)
+                        .collect(MoreCollectors.toImmutableSet());
+                Map<Id, ResolvedChannel> channelMap = resolvedChannelResolver.resolveChannelMap(channelIds);
+
                 List<ResolvedBroadcast> resolvedBroadcasts = broadcasts.stream()
-                        .map(broadcast -> ResolvedBroadcast.create(broadcast, resolveChannel(broadcast)))
+                        .map(broadcast -> ResolvedBroadcast.create(broadcast, channelMap.get(broadcast.getChannelId())))
                         .collect(MoreCollectors.toImmutableList());
 
                 writer.writeList(
@@ -103,57 +92,18 @@ public class UpcomingBroadcastsAnnotation extends OutputAnnotation<Content> {
                         ctxt
                 );
             } else {
+                Set<Id> channelIds = filteredBroadcasts.stream().map(Broadcast::getChannelId)
+                        .collect(MoreCollectors.toImmutableSet());
+                Map<Id, ResolvedChannel> channelMap = resolvedChannelResolver.resolveChannelMap(channelIds);
                 writer.writeList(
                         broadcastWriter,
-                        broadcastStream.map(broadcast -> ResolvedBroadcast.create(
-                                broadcast,
-                                resolveChannel(broadcast))
+                        filteredBroadcasts.stream().map(broadcast ->
+                                ResolvedBroadcast.create(broadcast, channelMap.get(broadcast.getChannelId()))
                         )
                                 .collect(MoreCollectors.toImmutableList()),
                         ctxt
                 );
             }
-        }
-    }
-
-    //TODO: Move resolution logic to query executor
-    private ResolvedChannel resolveChannel(Broadcast broadcast) {
-
-        try {
-            Channel channel = Futures.getChecked(
-                    channelResolver.resolveIds(
-                            ImmutableList.of(broadcast.getChannelId())
-                    ),
-                    IOException.class
-            )
-                    .getResources()
-                    .first()
-                    .orNull();
-
-            return ResolvedChannel.builder()
-                    .withChannel(channel)
-                    .withResolvedEquivalents(resolveEquivalents(channel.getSameAs()))
-                    .build();
-
-        } catch (IOException e) {
-            log.error("Failed to resolve channel: {}", broadcast.getChannelId(), e);
-            return null;
-        }
-
-    }
-
-    @Nullable
-    private Iterable<Channel> resolveEquivalents(Set<ChannelEquivRef> channelRefs) {
-        try {
-            if (channelRefs != null && !channelRefs.isEmpty()) {
-                Iterable<Id> ids = Iterables.transform(channelRefs, ResourceRef::getId);
-                return channelResolver.resolveIds(ids).get(1, TimeUnit.MINUTES).getResources();
-            }
-
-            return null;
-        } catch (InterruptedException | ExecutionException | TimeoutException e) {
-            log.error("Failed to resolve channel equivlents", e);
-            return null;
         }
     }
 }
