+++ atlas-api/src/main/java/org/atlasapi/query/QueryWebModule.java
@@ -37,6 +37,7 @@ import org.atlasapi.output.ChannelMerger;
 import org.atlasapi.output.EntityListWriter;
 import org.atlasapi.output.EntityWriter;
 import org.atlasapi.output.QueryResultWriter;
+import org.atlasapi.output.ResolvedChannelResolver;
 import org.atlasapi.output.ScrubbablesSegmentRelatedLinkMerger;
 import org.atlasapi.output.SegmentRelatedLinkMergingFetcher;
 import org.atlasapi.output.annotation.AggregatedBroadcastsAnnotation;
@@ -344,6 +345,11 @@ public class QueryWebModule {
     }
 
     @Bean
+    ResolvedChannelResolver resolvedChannelResolver() {
+        return new ResolvedChannelResolver(channelResolver);
+    }
+
+    @Bean
     ScheduleController v4ScheduleController() {
         EntityListWriter<ItemAndBroadcast> entryListWriter =
                 new ScheduleEntryListWriter(
@@ -353,7 +359,7 @@ public class QueryWebModule {
                                 "broadcast",
                                 idCodec()
                         ),
-                        channelResolver
+                        resolvedChannelResolver()
                 );
         ScheduleListWriter scheduleWriter = new ScheduleListWriter(
                 channelListWriter(),
@@ -1109,24 +1115,24 @@ public class QueryWebModule {
                 )
                 .register(
                         BROADCASTS,
-                        BroadcastsAnnotation.create(idCodec(), channelResolver),
+                        BroadcastsAnnotation.create(idCodec(), resolvedChannelResolver()),
                         commonImplied
                 )
                 .register(
                         ALL_MERGED_BROADCASTS,
-                        BroadcastsAnnotation.create(idCodec(), channelResolver),
+                        BroadcastsAnnotation.create(idCodec(), resolvedChannelResolver()),
                         commonImplied
                 )
                 .register(
                         ALL_BROADCASTS,
-                        BroadcastsAnnotation.create(idCodec(), channelResolver),
+                        BroadcastsAnnotation.create(idCodec(), resolvedChannelResolver()),
                         commonImplied
                 )
                 .register(
                         UPCOMING_BROADCASTS,
                         UpcomingBroadcastsAnnotation.create(
                                 idCodec(),
-                                channelResolver
+                                resolvedChannelResolver()
                         ),
                         commonImplied
                 )
@@ -1134,7 +1140,7 @@ public class QueryWebModule {
                         CURRENT_AND_FUTURE_BROADCASTS,
                         CurrentAndFutureBroadcastsAnnotation.create(
                                 idCodec(),
-                                channelResolver
+                                resolvedChannelResolver()
                         ),
                         commonImplied
                 )
@@ -1142,7 +1148,7 @@ public class QueryWebModule {
                         FIRST_BROADCASTS,
                         FirstBroadcastAnnotation.create(
                                 idCodec(),
-                                channelResolver
+                                resolvedChannelResolver()
                         ),
                         commonImplied
                 )
@@ -1151,7 +1157,7 @@ public class QueryWebModule {
                         NextBroadcastAnnotation.create(
                                 new SystemClock(),
                                 idCodec(),
-                                channelResolver
+                                resolvedChannelResolver()
                         ),
                         commonImplied
                 )
@@ -1185,7 +1191,7 @@ public class QueryWebModule {
                                                         idSummaryWriter
                                                 )
                                         ),
-                                        channelResolver
+                                        resolvedChannelResolver()
                                 )
                         ), commonImplied
                 )
